+++ .gitignore
@@ -1,3 +1,4 @@
 /node_modules
 debug.log
 build
+package-lock.json
