--- src/pages/Morse/index.js
@@ -1,4 +1,0 @@
-import MorseToText from './MorseToText'
-import TextToMorse from './TextToMorse'
-
-export { MorseToText, TextToMorse }
