--- src/pages/Text Transform/index.js
@@ -1,7 +1,0 @@
-import LowerCase from './LowerCase'
-import UpperCase from './UpperCase'
-import CamelCase from './CamelCase'
-import KebabCase from './KebabCase'
-import SnakeCase from './SnakeCase'
-
-export { LowerCase, UpperCase, CamelCase, KebabCase, SnakeCase }
