--- src/pages/URL/index.js
@@ -1,4 +1,0 @@
-import Encode from './Encode'
-import Decode from './Decode'
-
-export { Encode, Decode }
